+++ demisto_sdk/commands/init/contribution_converter.py
@@ -53,7 +53,7 @@ class ContributionConverter:
 
     def __init__(self, name: str = '', contribution: Union[str] = None, description: str = '', author: str = '',
                  gh_user: str = '', create_new: bool = True, pack_dir_name: Union[str] = None,
-                 base_dir: Union[str] = None):
+                 base_dir: Union[str] = None, no_pipenv: bool = False):
         """Initializes a ContributionConverter instance
 
         Note that when recieving a contribution that is an update to an existing pack that the values of 'name',
@@ -81,6 +81,7 @@ class ContributionConverter:
         self.gh_user = gh_user
         self.contrib_conversion_errs: List[str] = []
         self.create_new = create_new
+        self.no_pipenv = no_pipenv
         base_dir = base_dir or get_content_path()
         self.packs_dir_path = os.path.join(base_dir, 'Packs')
         if not os.path.isdir(self.packs_dir_path):
@@ -374,11 +375,11 @@ class ContributionConverter:
                         os.makedirs(output_dir, exist_ok=True)
                         extractor = Extractor(input=content_item_file_path, file_type=file_type, output=output_dir,
                                               no_readme=True, base_name=base_name,
-                                              no_auto_create_dir=(not autocreate_dir))
+                                              no_auto_create_dir=(not autocreate_dir), no_pipenv=self.no_pipenv)
 
                     else:
                         extractor = Extractor(input=content_item_file_path, file_type=file_type,
-                                              output=content_item_dir)
+                                              output=content_item_dir, no_pipenv=self.no_pipenv)
                     extractor.extract_to_package_format()
                 except Exception as e:
                     err_msg = f'Error occurred while trying to split the unified YAML "{content_item_file_path}" ' \