+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+
 # 1.2.17
 * Added a validation that the classifier of an integration exists.
 * Added a validation that the mapper of an integration exists.