+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+
 # 1.2.17
 * Added a validation that the taskid field and the id field under the task field are both from uuid format and contain the same value 
 * Added a validation that the classifier of an integration exists.