+++ CHANGELOG.md
@@ -18,6 +18,8 @@
 * Updated **generate-docs** command to take the parameters names in setup section from display field and to use additionalinfo field when exist.
 * Using the *verbose* argument in the **find-dependencies** command will now log to the console.
 * Improved the deprecated message validation required from integrations.
+* Fixed an issue in the **generate-docs** command where **Context Example** section was created when it was empty.
+
 
 # 1.2.16
 * Added allowed ignore errors to the *IDSetValidator*.