+++ CHANGELOG.md
@@ -1,16 +1,25 @@
 # Changelog
+
+
+# 1.2.17
 * Added a validation that the classifier of an integration exists.
 * Added a validation that the mapper of an integration exists.
 * Added a validation that the incident types of a classifier exist.
 * Added a validation that the incident types of a mapper exist.
 * Added support for *text* argument when running **demisto-sdk update-release-notes** on the ApiModules pack.
 * Added a validation for the minimal version of an indicator field of type grid.
+* Added new validation for incident and indicator fields in classifiers mappers and layouts exist in the content.
 * Added cache for get_remote_file to reducing failures from accessing the remote repo.
 * Fixed an issue in the **format** command where `_dev` or `_copy` suffixes weren't removed from the `id` of the given playbooks.
 * Playbook dependencies from incident and indicator fields are now marked as optional.
 * Mappers dependencies from incident types and incident fields are now marked as optional.
 * Classifier dependencies from incident types are now marked as optional.
 * Updated **demisto-sdk init** command to no longer create `created` field in pack_metadata file
+* Updated **generate-docs** command to take the parameters names in setup section from display field and to use additionalinfo field when exist.
+* Using the *verbose* argument in the **find-dependencies** command will now log to the console.
+* Improved the deprecated message validation required from integrations.
+* Fixed an issue in the **generate-docs** command where **Context Example** section was created when it was empty.
+
 
 # 1.2.16
 * Added allowed ignore errors to the *IDSetValidator*.